♻️ Refactor: 모임 ApiService 분리
Related to: #318
