+++ app/src/main/java/com/mongmong/namo/data/remote/group/GroupScheduleApiService.kt
@@ -1,5 +1,7 @@
 package com.mongmong.namo.data.remote.group
 
+import com.mongmong.namo.data.dto.GetMoimDetailResponse
+import com.mongmong.namo.data.dto.GetMoimResponse
 import com.mongmong.namo.domain.model.group.AddMoimScheduleRequestBody
 import com.mongmong.namo.domain.model.group.AddMoimScheduleResponse
 import com.mongmong.namo.domain.model.group.EditMoimScheduleRequestBody
@@ -14,6 +16,16 @@ import retrofit2.http.POST
 import retrofit2.http.Path
 
 interface GroupScheduleApiService {
+    // 모임 일정 목록 조회
+    @GET("schedules/meeting")
+    suspend fun getAllMoimSchedule(): GetMoimResponse
+
+    // 모임 일정 상세 조회
+    @GET("schedules/meeting/{meetingScheduleId}")
+    suspend fun getMoimScheduleDetail(
+        @Path("meetingScheduleId") moimScheduleId: Long
+    ) : GetMoimDetailResponse
+
     // 그룹의 모든 일정 조회
     @GET("group/schedules/{groupId}/all")
     suspend fun getAllMoimSchedule(
